rfac: Move loggingMiddleware
