--- internal/server/logging_test.go
@@ -1,40 +1,0 @@
-// Copyright 2026 Jip de Beer (Jip-Hop) and ojster contributers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package server
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-)
-
-func TestLoggingMiddleware(t *testing.T) {
-	called := false
-	h := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		called = true
-		w.WriteHeader(http.StatusTeapot)
-	})
-	mw := loggingMiddleware(h)
-
-	req := httptest.NewRequest("GET", "/x", nil)
-	rec := httptest.NewRecorder()
-
-	mw.ServeHTTP(rec, req)
-
-	if !called {
-		t.Fatalf("handler not called")
-	}
-	ExpectStatus(t, rec, http.StatusTeapot)
-}
